refactor: pull base annotation encoder into models
